DISPATCH-1322: Edge router test drops a disposition
This patch has a stand-alone reproducer for exposing the problem
in DISPATCH-1322:

 * A shell script runs two routers
 * A client program runs the test

The client program runs the test up to N times looking the failure.
The test always stops on failure instead of completing N passes:
  this keeps the log files from getting too huge.
The client keeps a uS-timestamped log for coordinating the test
  actions with the router logs.
There is a clean.sh script to ready the logs and output between runs.

To repeatedly run the test looking for a failure use:

  killall qdrouterd ; ./clean.sh ; ./run-routers.sh ; sleep 5 ; ./client.py