DISAPTCH-1749: Improve test framwork router.wait_ready function
This patch adds an optional new check for testing that a router is ready.
By default this check is enabled.

The check is to locate the message "SERVER (notice) Operational" in a
log file and to time out if this message is never seen.

This check is added in addition to the existing port and connector probes.